qvector: makes qv_insert/qv_append safer.
The current implementation was not resilient in case of insertion of a
copy of an element that is already present in the qv. For example:

 qv_append(t, v, *qv_last(t, v))

Because of the absence of a sequence point in the implementation, there
was no evaluation order between the resize of the vector and the
qv_last(), which led to potential crashes.

Fix this by ensuring the added value is computed before the rest of the
process. This behavior is more "function-like" in that parameter are now
evaluated before the body of the function (argument passed by value).

Change-Id: Ied55eec2a1c00fdd80906816f03e186f43676e92
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>