time: avoid overflow in time_fmt_iso8601_msec
time_fmt_iso8601_msec was wrongly used in lms and mcms, we passed
microseconds instead of milliseconds. This actually caused a buffer
overflow, but since we use the sb variant, we never detected it.

Now we check the range of msec, the return value of gmtime_r (which can
fail if tm.tm_year does not fit on an int), and we use snprintf to avoid
buffer overflow and we check the return value to see if it was truncated.

If any error occurs, we panic, it's the responsibility of the caller to
pass a valid timestamp (the current use cases always pass the current time
as far as I can tell).

Change-Id: Ibb66f9bea0c12c69eb058a4b3bcffa89cf9fb841