mem-ring: delay deletion of mem-ring if they still contain frames.
Since the deallocation often takes place in another thread than the one
that owns the allocator. As a consequence, sometimes we could release
frames after the owning thread deleted the ring (for example when the
r_pool() get deleted at the shutdown of a thread while the main thread
didn't release all the frames, yet).

To avoid this, have the same mechanism as the mem-fifo: if the ring get
deleted while still having active frames, it will not be deleted
immediately but when the last frame is deleted.

Change-Id: Ifa70f0daaa0481407e93a354d1a94f290f97e691
Closes: #30781
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>