qvector: fix leak on mp_qv_init with size 0
When a qv was initialized with NULL as memory pool and with a size of 0,
and then wipped, free was not called on the pointer returned by a malloc
of size zero, resulting in a leak (according to asan).

This problem is fixed and a test that triggered the leak is added.

Change-Id: I20f053e367a207af1d7b391d3d7c7f45f0419f9d