log: Use __LOGGER_LOG to implement logger_log.
Before this commit, logger_log was implemented directly by calling
__logger_log.

The problem is that __logger_log (which calls logger_vlog) is using
logger->full_name to get the name of the used logger without checking its
conf_gen.
As a consequence, logger_log called on a logger which was never used
before is the same as emitting a log without logger.

This is a regression of 52a22412, where logger_vlog was reworked.

To fix that, simply implement logger_log using __LOGGER_LOG, which calls
logger_has_level which itself refreshes the logger if needed.

Also add asserts in logger_vlog and __logger_start to be sure that the
logger is up-to-date, which is supposed to be the case.

Change-Id: Iea146ae722cab1a0859c18b574ca4aa214bbfcb1