container/qvector: add mp_XXX_new() helpers for qh/qm/qv
The reason why these macros are added is that we want to ease the qh/qm/qv
manipulation in blocks.

For example, it can simplify the code for synchronous block-based
functions:

    t_scope;
    qm_t(foo) _foo;
    qm_t(foo) *foo = &_foo;

    t_qm_init(foo, foo, sz);

    do_something_with_blk(^() {
        qm_add(foo, foo, [...]);
    });

Instead, these new macros allow to write:

    t_scope;
    qm_t(foo) *foo = t_qm_new(foo, sz);

    do_something_with_blk(^() {
        qm_add(foo, foo, [...]);
    });

Of course, it implies one allocation in the t_stack versus one slot taken
in the stack, which is slightly more expensive, but expected to be
negligible considering the context.

This patch also add Z tests to ensure that the macros can compile
properly.

Warnings:
* There is no "__must_check__" for these macros;
* No mp_XXX_delete() helper is added for now, so the new helpers are
  designed to be used exclusively with allocators that does not implement
  a 'free' function, like the stack pool or the ring pool.

Change-Id: Ic73d3b4d23a741643b83ccb84d7d1be0a4d6638a