Fix e_traced & co when called with no-constant level or name.
e_name_is_traced was evaluating its result the first time it was called
for a specific e_trace line, caching it and then returning the cached
result in any following call.

The problem is that the cached result was never re-evaluated, whatever the
given level -> only the level given at the first call was taken into
account.
This works fine with constant levels, but is buggy with non-constant ones
(and we have some in our code-base).

To fix that, simply not cache the result when the level is not a constant
(and do the same for the name, because we have the same bug).

Change-Id: I4a6009999dec4ab7614e74194bd89c036424d6d1