pstream: unify behavior of parsing uint64_t strings.
The functions lstr_to_uint64() and ps_get_ull_ext() are used to parse a
string to an unsigned integer.

The goal of this commit is to unify the behavior of both these functions:
return an error when the string is a signed number.

lstr_to_uint64() had already been patched and has the correct behavior,
but ps_get_ull_ext() is accepting negative values.

These functions are calling memtoull_ext and memtoullp (that are wrappers
around strtoull()). These wrappers have been modified to throw an error
(ERANGE) in case of negative values.

The fix in lstr_to_uint64() is now useless and has been removed.

To sum up, the following public functions are throwing an error in case
of signed number:
- memtoullp (new)
- memtoull_ext (new)
- strtoull_ext (new)
- ps_get_ull_ext (new)
- lstr_to_uint64

It follows this patch: https://git.corp/r/#/c/38462/ that was
reverted because it was failing with the compilation default mode (and
not with asan mode). The output value was checked if errno was not equal
to EINVAL. The output value should be checked when errno is not set.

Change-Id: I92da9259ac468ef98d9a93135a00d30457d34d7a