log: Call logger_has_level in e_* functions.
The assert added in 379e99a17e2fe was triggered in the C tests of
platform on centos-6:
 https://git.corp/buildbot/builders/review-platform/builds/16384

This is because the good old e_* functions were calling directly
logger_vlog, so the same problem as explained in 379e99a17e2fe
happened when using them on the root logger.

To fix that, call logger_has_level in these functions.

Change-Id: I22b6fb721aae7b8218d16955e85d505e9d710128