thr-job: Introduce THR_MAX_PARALLELISM environment variable.
This environment variable allows to set the maximum value of
thr_parallelism_g (which in any case cannot be greater than the number of
CPU).
This determines the number of threads that are launched in thr_initialize.

This patch was made to be able to set a reasonable value for this
parameter on the external buildbot machines having so much CPU that in
many cases threads creation are failing, as observed here:

 https://git.corp/buildbot/builders/review-platform/builds/16772/steps/test/logs/stdio

Change-Id: Iab1a4d14c0e0423b72d91c80ad00f0654bc9a940