core-havege: ha_rand_range did not support a [INT_MIN:INT_MAX]
When calling ha_rand_range with a range [INT_MIN:INT_MAX] the results
were always INT_MIN.

This was due to an int overflow in ha_rand_range.

A test has been added to check that the bug has correctly been fix, but
does not check the correctness of the function ha_rand_range.
Such a test should be written

Change-Id: I9a6035419c910d2e85bc8b9a9995b024da558c44