ichannel: fix regression on writing failures
The regression was introduced by
I3c86b68ddfcf01cae182252fa13f4e2af836e47c. When a write interleaved with
the reading loop fails we want to finished the loop but we still need to
close the connection at the end of the loop. Otherwise it would leave the
channel blocked.

Closes: #31938
Change-Id: I20745244bb83b352ea17b4fdba906ce3383beb00