file-bin: Fix t_file_bin_get_last_records.
This function was buggy in case the number of required records was greater
than the actual number of records in the file:
* it was looping over and over on the first slot until the number of
  wanted records was reached.
* in the last read slot, only the first "count" entries were read, which
  was an error because this is a reverse run: the last entry of the slot
  is the first entry to add in the output vector.

This was not seen before because the tests were giving the exact number of
expected records, which was a quite bad idea.

Moreover, we were possibly missing the last record for a slot if the end
of the slot was only containing the record header and no data.

+ fix the minimum size check for a slot. A slot should be able to contains
file header + slot header + record header.

+ fix when trying to get a record from an empty last slot of a file

+ fix when reading the first slot, which has a different size (because of
the file header)

Change-Id: I271926725efa7c24c96b788ce4ac4b267129a267
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>