python: save the parse cbk exception messages in a variable
Whenever an exception was raised in a callback the message was not printed
so one had no clue about the reason of the exception.

We now catch the exception and send it back to the query_end callback.
One has the print/log the error message from the query_end cb to see the
exception message

Change-Id: I775808ad7ef93852c365c447532465686befc10d
Refs: #30538
Signed-off-by: Nicolas Angevelle <nicolas.angevelle@intersec.com>