ichannel: add a mechanism to add tracing on a per-query basis.
A new flag is added to the ic_msg_t object that instructs the ichannel
layer to enable tracing for the corresponding query. Once done, traces are
printed in various step on the process both locally and remotely. The
remote part is achieved using a new flag in the ichannel protocol:
IC_MSG_IS_TRACED that is sent alongside traced query.

This let the user identify when queries are sent, when they are received
by the remote end and what reply is given (and received).

Many coding rules issues have been fixed at the same time.

Change-Id: I89e96d0829c60f481962a2e926f93c5f0ba5e8e7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>