ichannel: allow activation of tracing through an environment variable.
The environment IC_TRACE can be set. It should then be a space-separated list of interfaces'
fullnames or RPC fullnames.

Every listed RPC/interface is then automatically traced when an ic_query
is performed.

It should be noted that this is activated in development builds only in
order to avoid possible side effects of the tracing mechanism on
production.

Morever, in order to see the logs, you must ensure that the ic/tracing
logger is activated:

 IC_TRACE="qkv.Base qkv.Repl.push" IS_DEBUG="+ic/tracing:0" ./zchk-cluster

Change-Id: I85ba57af0a9b046f1b50866f9f78cd25e3d8e305
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>