file-bin: Make file_bin_get_next_record robust to incomplete files.
file_bin_get_next_record was not written so that it can support parsing
incomplete files (ie. without the end).
This is a common use-case which is needed for example by B-On (the logd
writes a file which is read at the same time by the enabler action).

This commit patches this function to support it. This mostly consists in
jumping to the next slot in case of parsing error only if there is enough
space in the file (this was not checked before), and get back to the
previous position otherwise.

Also detect corrupted record length in some cases.

Note that file_bin_for_each_entry was modified because "the file is
finished" cannot be a valid stop condition (because if the file is
incomplete, we never reach the end).

Change-Id: Id3d2a182c1ab1e96afd0ddfd6902831e7d8d6bea