Build: checks, split ZFile dir and test path
The web team wishes to order tests suites files in some sub-directories.
But testem must be run in main directory where is located testem.mustache.

This patch splits the ZFile directory and the test path.

So, now the rule is to create testem.mustache and ZFile
in the same directory.

Change-Id: I3db9494de9818d93d01a02e97fc5f90ac8b92428
Closes: #32663