lib-common: possibility to timeout and/or cancel ic queries
This patch adds two great features on the ichannel library.

The first one is the possibility to cancel a query. When a query is tagged
as canceled, it will be skipped when sending the inqueued queries (and it
will be properly deleted). If the query is already sent and is
"synchronous", we will just drop the answer.

The second feature is the possibility to add a timeout on a ic_msg_t.
When sending a query with a msg that has a timeout configured, the query
will now fail if it is not entirely processed before it times out. On such
cases, the callback of the query is called and the status will be
IC_MSG_TIMEDOUT.

Change-Id: I1c293ebc4671ca1032168ec67698a69485634144
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>