wah: fix issue when adding more than 64G entries at once.
Run in WAH store the number of words in the run in 31 bits. As a
consequence, a run cannot contain more than 64G bits. This was not
properly handled in the code that add data in the QTS. We now add code
that ensure that we dispatch the data in several runs if the number of
bits exceed the limit of the run.

Change-Id: Ie803000c6a76fd1cdb13b4886b3f592d1090584e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>