lib-common: fix ic_abort
Problem
~~~~~~~

On rare cases the callback of a query is modifying the message queue (with
a use of ic_mark_disconnect in shutdown procedures for example). This is
problematic on ic_abort because we try to extract the ic_msg from the
queue after the callback (thanks to ic_msg_take_and_delete).

Solution
~~~~~~~~

On ic_abort, get rid of ic_msg_take_and_delete and extract the message
from the queue first, then call the callback.

Change-Id: I33550046264e6233451bd1b203f45ba22831c4ec
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>