container/qvector: add documentation for qv_splice()
With three integer parameters and ambiguous parameters names, qv_splice()
is definitely not as straightforward as most of the other macros and
functions available for qvectors, and really needs some documentation.

Note: the parameters names are questionable IMO. For example, I have no
idea what "dl" stands for and the only thing I can think of is "deletion
length", but "dl" is the insertion length, the deletion length being given
by the parameter "l".  The parameters names may be reworked in another
patch.

Change-Id: Iedcc72d28cc8d217dc6bbf2bece49c525becc83f
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>