qps-bitmap: fix missing dereference of the root in qps_bitmap_reset().
We use the map->root->nullable flag in order to determine if the 0 is
stored or not in the bitmap but we didn't deref the root_cache handle
before using the pointer. As a consequence, if the QPS has been garbage
collected, map->root could be outdated and the access caused a crash.

Change-Id: I085b3831a50535b7620c9802c300d200e728dc7d
Closes: #33079
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>