container-qvector: rework qv_extend() in order to fix a rare mem error
When extending a vector with itself, the way qv_extend() is implemented
breaks the number of elements to copy into the new created slots.

Rework the implementation so we can write:

    qv_extend(i32, &vec, &vec);

Note: we don't call qv_growlen() into the p_copy() in order to make sure
that qv_growlen() is called *before* "__tab->tab" is evaluated (the value
of "__tab->tab" can change in case of realloc).

Also add a zchk for that particular case.

Change-Id: I2f04bae9066aa00ff1eb424625d236a249230fd2