file-log: trigger CLOSE event after closing the last file
The CLOSE event could be triggered and handled in the callback before the
file is flushed and closed, causing possible race conditions:
In my case, my callback does a rename() of the file to move it in a
data-collector incoming directory. I may be a bit paranoid, but I fear
that the data-collector could read the moved file before the flush is done
in log_file_close and then process the file only partially.

Change-Id: I3721ef7ac76c154b3fe4c8a4f50375aba348cc85