core-obj: fix race in creation of obj_class_t objects
When creating a new object, the function pfx##_class() is called, which
lazily creates the class information. However, if several threads call
this function at the same time, a race can occur where we set the
"res->super" information before having set the rest of the object, so the
object can be used in an inconsistent state.

To make sure everything is perfect from every thread, several memory
fences should be used in an acquire/release mode. A spin lock is also
added so that only one thread creates the object.

This fixes a fail in the qmr tests (mostly zchk-sc.py, could happen in
zchk too) where all the threads were creating the same object at the
same time in the map step.

Closes: #25088
Change-Id: I180957b45705d6abdab88e3ea1b0ac50c163858e