iopc: keyword `local` for class.
If the class parent is defined local using the keyword local, the class
son needs to be in the same package than the class parent. If not, the
error "as the parent class `%s` of class `%s` is `local`, both classes
need to be in the same package" will occur.

The keyword isn't inherited.

Closes: #33504
Change-Id: I24c77b7d400168a48e3b899273e99a2aedfc15ec