iop-rpc-channel: Better handling of invalid encoding.
As observed in #33661, having the "invalid encoding" warning in the logs
could be very useful in production too, so let's turn it into a warning.

The only good reason I see to hide it is to avoid to be flooded by a
malicious client. To avoid that, close the connection when this happens.

Closes: #33661
Change-Id: I83c4599aa7c33d6f18437d352743bbf46142daf7