core-obj: really fix race in OBJ_VTABLE
Hopefully, the race is now completely eradicated. The issue was with the
memcpy instruction which modified the conditions and allowed a non
complete object to escape the function.

The pfx##_class() is rewritten to be much more robust, using an atomic
type and a temporary variable to avoid any side effect from one thread to
trinkle down in other threads without it being explicit.

TSAN doesn't complain anymore about this function, so this should now be
thread-safe.

This fixes another random bug in QMR, where the lib-sum test return 9900
instead of 10000.

Change-Id: I45cea40ff9c215285744ae6851d15db945902fa9