wah: fix highly unbalanced wah_and.
When a wah_and was called with bitmaps that have more than 128G bits of
difference in length, we ended up with an infinite loop because the
remaining number of words was casted to a 32bits and overflowed.

Fix this by capping the max number of words that can be stored in a single
wah-run (we could also have choose to switch to a 64bits integer, but
since we cannot consume more than WAH_MAX_WORDS_IN_RUN in a row,
capping is safer).

Change-Id: Iff085980df03cfddd9bccfde45a29e4537923e09
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>