force TZ to Europe/Paris in Z tests
All the tests using dates and timestamps assumed we were in the
Europe/Paris timezone, since that's how all our machines are setup.

Running with a different timezone is clearly broken, for instance, try the
following in platform/platform:
TZ=America/New_York make check

By forcing the timezone, we can allow our tests to work on setups with a
different timezone.

Note:
* this was already done in a Z_TEST in lib-inet/sms-gsm.c
* python tests are also broken, and will probably need to add an optional
  timezone parameter in platform time_cfg

Change-Id: I8fbdebb98f35e1220ca5f2c215e3515476a68439