core: update the expect() macro to ensure its must-check'ness
It is probably due to an update of the compiler, but recently expect has
not been a must_check anymore. Removing the likely of must_check restores
this behavior.
Plus, expect remains to be likely to be true since it calls e_expect,
which is unlikely to __assert_fail and returns true otherwise.

Change-Id: Ifb3a5c7a56cb64b0fd79109396315997c7ca088f