module: add a wrapper around the opaque structure module_method_t
The idea is to pre-compute the calls to a method.

In order to do so a qv_t() of callbacks must be stored in the
module_method_t.

As the module_method_t is exposed in the header we cannot use directly
vectors in it, so a module_method_impl_t has been created.

Change-Id: Ie2313d96d5edea538c522d33c22736953db8c7e2