module: compute method callback ordering only when the ordering might change
Previously the run method was recomputing each time the order in which the
method callbacks should be called.

The ordering of the methods will change only when new modules get release
or require, so there absolutely no need of recomputing each time the
order.

This method callbacks ordering is now computed only when a module get
release or require by storing the ordering in a qv for each method.

Change-Id: I5d053bd3e4e2b6a56251cfabdd1d78e44aea4ae0