container/qvector: make qv_init_static() more strict
The helper 'qv_init_static()' allows to initialize a vector with data not
owned by the qvector, as a consequence, it is supposed to be a little more
permissive than the other initializers. The problem being: excessive
permissiveness: the type of the pointer given as an argument could be
anything.

Change-Id: If13a1e6e0c793679390a7b414d1c8ce541ffda3d