str-iprintf: add ability to extend the %*p formatter-set.
Currently, str-iprintf supports %*pM, %*px and %*pX. This patch adds the
ability to register custom formatter for custom modifiers.

A modifier must be a single ascii character, and not conflict with any
other formatter. The formatter must be able to able to handle both file
streams and buffer.

Internally, the support for existing modifiers has been ported to the new
format, only %*pM has a special treatment since it's used very often.

Change-Id: I36992a2503c1b1ea4b78cea769dc00e829d5adac
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>