z: Also check name of tests registered with Z_TEST.
Unlike what I said in I5e4f194f4 (which was what I though when writing
it), the first argument given to the Z_TEST macro is not used to form a C
variable name, so this is possible to put spaces in it. Also check them to
avoid that.

Group names registered with Z_GROUP_EXPORT are not concerned.

Change-Id: I205f4a7fa507147b5c0ce8e5c09a1d08b6ebb055