str-buf: Always have final \n in sb_getline result.
Problem:
~~~~~~~~
With the previous code, the sb filled by sb_getline was always terminated
by a '\n', except maybe for the last line if the file is not terminated by
a '\n'.

This is an issue because it had to be taken into account by the callers,
complicating the parsing of the result.

Solution:
~~~~~~~~~
Be sure the filled sb is always terminated by a '\n' when a positive
result is returned.

Change-Id: I6a96b8c00670c451c38b6f79dc847a532c95db99