httpd: account queries only when they get parsed
A query was accounted directly after its HTTP headers reception. It wasn't
a problem unless we set the server pipeline depth to 1 and the client
a bigger value. It led to a staled connection with this flow:

     client                            server
   send-query(1) ---------->
   send-query(2) ---------->  read(HDR1 + BODY1 + HDR2)
                              queries-count(2)
                              2 >= pipeline-depth => remove POLLIN
                              handle(query1)
                 <----------  write(query1-response)
                              queries-count(1)
                              never restore POLLIN since 1 still >=
                              pipeline-depth and never read(BODY2)…

The trivial fix here is to only account the query after it gets fully
parsed. There is a several ways to fix it but this one seemed to be the
simplest.

Change-Id: Ic992feb1d39063460938be057e1cdb6bb28f80e4