httpd: avoid to drop our last responses before closing the connection
In several TCP stack, like the Linux one, when we close a connection if
there was some pending data in the reading buffer we will emit a RST
instead of a FIN.

A notable consequence is when we decide to close our connection because we
have reached the maximum number of queries (or whatever reason), if the
client has continued to send some data it will receive a RST and our last
responses will be dropped. From the client viewpoint the last requests
will be aborted and thus should be remitted. Which will lead to the
reception of duplicated requests on the server.

The clean way (the only way) to fix this would be to half-closing our TCP
connection (the writing side) and wait until the client closes the
connection itself. But this behavior has several security implications and
we don't want to trust the client. So we use a well-known (and used) TCP
hack. We call shutdown(fd, SHUT_WR) just before our close() call.

With this hack, we will still emit a TCP RST, _but_ the shutdown(fd,
SHUT_WR) will place a FIN flag on the last written packet and for
a flush of the writing buffer. Thus the client will correctly received all
our responses before the brutal RST (the FIN flag will preserve our
packets from dropping).

More details here:

http://blog.netherlabs.nl/articles/2009/01/18/the-ultimate-so_linger-page-or-why-is-my-tcp-not-reliable
http://blog.olivierlanglois.net/index.php/2010/02/06/tcp_rst_flag_subtleties

Change-Id: Id5d8e4b942b152c3cd134ee51a8506a3c5f47f82