core-macros: add nullability extension support.
This (future) extension from clang looks very handy in that is allows the
expression of the nullability of the parameters/return value of a function
in a very explicity way. E.g.:

 nullable char *strchr(nonnull const char *s, int c);

This is far more explicit than the __attribute__((nonnull())) and much
more robust since the qualifier is placed alongside the variable.

Note that the qualifiers applies to the variable, not the type.

See the RFC on clang-ML for details:
http://thread.gmane.org/gmane.comp.compilers.clang.devel/41568

Note that in the current state, these qualifiers won't do anything but act
as API contracts/documentation, which is pretty useful.

The main implication of this patch is that we must use __nonnull__ for
attributes instead of just nonnull to avoid macro conflicts. BTW, we use
nonnull instead of __nonnull because the libc defines __nonnull as an
alias to __attribute__(__nonnull__)... which is a shame!

Change-Id: I36dcc64931d580017d3f3e46559b0eee935105a4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>