iopc-dso: always rebuild dso file
the dso does not only depend on the iop file (event.iop in current
platform use case), but also on all iop dependencies given in env.

In the platform use case, base.iop is hardcoded in the binary, but we have
currently no way to track its version.

Since rebuilding the dso at startup is not too costly, let's do that.

Change-Id: I7cfac2556f2893ef9479e2432d422e6acf2b2568
Closes: #35188