iop: don't define per-structure base helpers.
We now define helpers à-la 'iop_new()' for init/dup/copy/equals/check/...
instead of defining new functions for each structure/class/union.

The main purpose of this patch is to reduce the compilation time of the
repository by removing functions that get generated everywhere but are
actually used in a very local call points.

Next patches will remove helpers for enums and packers as well, but we
keep this in several patches in order to ease the update in the other
repositories.

Change-Id: I91aa80760d460ad3623f444c2cdc4dd4878e52db
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>