iop: remove per-struct/union packers/unpackers.
This removes a lot of generated function.

Note that this patch does not provide helpers as done in the previous
ones. This is done on-purpose since a deeper refactoring the
packing/unpacking API will be required in order to improves the
consistency of the API and the ease of use.

Change-Id: Idbad562738052d003d5fd4f035d15b6fc23ec07e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>