lib-common: don't keep generated IOP file in repository.
Those files are in the repository because they are needed in order to
bootstrap the tools repository. However, only the master branch is used by
the tools repository, so keeping those files in older branches is useless.

Moreover, keeping those files is weak and dangerous since we use a single
iopc whose exact output may change (while still generating backward
compatible code). As a consequence, we end up having the repository marked
as "dirty" on old branches when only generated files change (leading to
rejected deliveries).

As a consequence, only the master branch will keep those files. The master
branch is now even more considered as unsuitable for delivery since we
don't provide a clean way to ensure it won't lead to dirty deliveries.
When a new branches are creating a new lib-common branch, one must kill
the generated files from that new branch.

Note: the change is done on 2011.2 since earlier branches since it's the
older branch with a separate iopc.

Refs: #21903
Change-Id: I76f52e736194f93d6091e7376495ce569123e2bf
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>