iop[rpc]: fix header replacement when using a dynproxy
We used to check if the original header was changed by the dynproxy with
ic__hdr__equals(). But the IOP-header is generally changed in place and
thus we can't see the change with a comparison.

So we always have to re-pack the given header when the dynproxy returns
a header and we have already one.

Change-Id: I1de98abd4764fe6bfc81b2ccfc9f84a5e436c9d9