iop[channel]: document that dynproxy may return memory from the t_pool()
When we want to return a complex header from the dynproxy we often need to
allocate it and it becomes hard to track this memory. The IChannel library
can ensure that we are safe to return some memory from the t_pool() and
simplify things.

So we document it in order to preserve this behavior.

Change-Id: I95cc80e81e36ecb36ba2fe287d18999cd83854a4