log: introduce logger_exit
In some cases, we want to print a message (in both log and console), and
exit immediately. It is not a crash, only an error that can happen in
production environment (so we cannot use logger_fatal/logger_panic).

logger_exit does now the job: we print the message as an error (useful for
developers), we send the message to the syslog (useful for administrators)
and we call the _exit function. It does not call the destructor, it only
close the opened files and terminate the process as soon as possible.

Refs: #35758
Change-Id: I5c26a8ed5802f83fe2e23a06b6ca7913ffc28eaf