www/ci: Add www-coverage target
Basically we:
* copy the whole repo to .build-coverage-${hostname}
* replace meaningful js files by the instrumented ones
* Launch the proxy which receives the coverage objects and write them on
disk
* Run the test in .build-coverage-${hostname}
* Create the html and lcov report from all the collected coverage*json
files

There is a few tricks:
* the --baseline-file option creates a coverage object even for files which
are untouched by tests. We need a unique --save-baseline name for each
instrumented directory.
* --complete-copy dereference links and we have recursive loops and broken
links so we use "cp -r -a" plus the "--complete-copy" to bypass it
* We do not want coverage on external files, yet they are needed, so we
copy them afterward
* Writing the instrumented files in the .build-coverage-${hostname}
requires to remove symlink first or they'll be written in the source dir.

The make www-check line seems to launch tests on both
.build-coverage-${hostname} and source dir. It is not impacting the
coverage but is is just twice longer to complete the coverage.

Change-Id: I561b9beb5886b332a525446373c360954ecc7dad