str: add a utility function to pretty print a table.
In some case (dumper, ctl, ...) we want to pretty-print a table. Writing a
proper printer is not quite interesting. In order to avoid spending more
time doing this, have a pretty printer in the lib-common. Also introduce a
new str-buf-pp module to store this kind of functions (I plan to do the
same for tree pretty printing).

Change-Id: Ifa17541cc595426012503e7f91191338c30b5868
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>