www/coverage: Exclude cache
We don't want to add instrumented file in cache directory.

The cache dir from the source dir contains symlinks to the actual source
files. "istanbul instrument" is then  copying and instrumenting the
referenced file in the cache dir of the build dir. That creates two
problems:
* when the referenced file does not exists istanbul fails
* when it does exists it will create two different files and add coverage
for both (instead of merging the two files)

Change-Id: Ia7790437cee1e086e19edab2691f7dfa9893c37c