iop-rpc-channel: Fix sending of fd with sendmsg.
Problem:
~~~~~~~~
There was a really bad bug in the case sendmsg was called several times in
ic_write_seq, and that a fd was sent with the first call, and no fd should
be sent with the second one. This can happen if at least two "big"
messages are queued, one with a fd and the other without.

In that case, the second call to sendmsg was sending the fd again... which
was already closed because the associated message was deleted between the
two sendmsg calls. This caused a sendmsg failure (Bad file descriptor),
and then the destruction of the ichannel.

This is because of a bug in sendmsg: if no fd needs to be sent, nothing is
done in this function, which should instead reset the fds used in the
previous call.

Solution:
~~~~~~~~~
Simply delete this stupid check, the case where fdc == 0 is handled
properly 4 lines below.

Closes: #36223
Change-Id: I6d27c3fc80caac4a3d0872e38454e908fa6ca352