iop-mib: insertion of the conformance groups
Insertion of the conformance groups. For now there are two conformance
groups, one for the notifications and the other one for the objects.
They just list the different objects and notifications of the MIB.

Two buffers are used because we want ,for each field and rpc, to at the
same time write information for the MIB on a buffer and register the name
of the field and rpc for the conformance groups. The tricky point is that
the conformance group is ahead to the fields and rpcs information on the
MIB. This is why there are two MIBs.

TODO : Create deprecated groups

Refs: #34465

Change-Id: I6bd440ad1ac02ef25e1b585191e3d60701e82367