Build: improve checks on website init
When the machine/buildbot is not right configured we used a "exit 1"
which was not trigged by our build system.
And the buildbot occurred a wrong +1.

So, we use return instead exit to catch the error code.
Also, we do not initialize website if the product/module does not have
front end.

Change-Id: Icae06147939b7a912d221c583dc074ed2e03eb09