file-log: Do not ignore file_close in log_file_rotate.
I don't know why we were ignoring this, but keep going when the closing
the file failed looks dangerous and could hide some problems.

When trying to reproduce #36340 I was in the situation where file_write
did not fail while the file system was full. This patch fixes this
(because file_close performs a file_sync, which fails if the fs is full).

Change-Id: I48fa747ca83013dc000e496af5bd84d7a7f362ef
Refs: #36340