iop: fix iop_get_field for classes, references and optionals.
In the loop, we never checked if ptr is not NULL, this can lead to a
segfault, for example when using an optional that is not set.

When we wanted to get a class, reference or optional field, a
dereferencing was done at the end, so we received the value of the field
instead of the field itself.
For example, if I have a structure A that contains a class B in a field b,
before, when getting b, the type of out_ptr was a B * and I was not able
to point the field b to another instance of B. Now, out_ptr contains a
B **.
This behaviour is consistent with iop_value_from_field,
iop_value_to_field and platform_parameter_value_to_field.

Also add new tests and check every out_ptr values in tests.

Change-Id: I3578651d65619589e526b332b2c004b8fa7c200c