thr-job: allow custom handler for SIGSEGV and SIGBUS in threads.
In the loop that forks the thread, we ensured we blocked all the signals.
Since the threads inherits the sigmask of their parent, we completely
blocked all signals handling in those threads. For some signals this is a
desirable behavior (for example, we would prefer not being annoyed by
SIGCHILD signals in secondary threads) since their handling can be delayed
(and actually waits for the main thread to be ready to handle them), but
for other signals like SIGSEGV or SIGBUS that must be handled
synchronously by the thread that triggered them. When blocked, the default
behavior was used leading to the sighandlers to be completely bypassed
(including our ".debug" dumper)

Change-Id: I8a131b9e94bb5e06636cd830ce9c7d46bc76d7ef