el: get rid of the need to mask SIGCHLD when forking.
We used to have to use the following pattern:

 sigset_t set;

 sigemptyset(&set);
 sigaddset(&set, SIGCHLD);
 pthread_sigmask(SIG_BLOCK, &set, NULL);
 if (fork() == 0) {
      ...
 }
 el_child_register();
 pthread_sigmask(SIG_UNBLOCK, &set, NULL);

This was required because el_child_register() is triggered by the SIGCHLD
signal, as a consequence, we must be sure we don't receive the SIGCHLD for
the newly forked process before it has been registered in the event loop.

In order to work this around, el_child_register() now blocks the SIGCHLD
internally and, in order to cover the case where the child process already
died, it starts with a waitpid() on this child. That way, we are sure that
either the child died before the signal, in which case the waitpid() catch
it, or the corresponding SIGCHLD will be received after the blocking of
the signal has been released, that is after the el has been registered.

In case the process is caught it is added in the before list, in order to
ensure el_child_register() will still return a valid el_t and that the
callback is still called asynchronously from the event loop. A remaining
downside, though, is that, in some very specific and unlikely cases, the
pid may have time to be reused before the callback is called, which could
lead the application to wait for two processes with the same PID.

Change-Id: I05f91fc61bd6bf8785bd546a742c45b55e90d2ae
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>